Add summary for Jupyter files
Decouple myst_nb dependency from the main code by using entry points.
This allows the use of the extension without myst_nb being installed.
